x86: fix two i686-only bugs — per-thread TLS init and a return-path #GP (v0.64.14)
1) Per-thread ELF TLS (__thread) on i686 read 0 instead of its .tdata image,
   so M29's thread-local-bleed test (and any threaded i686 program) failed.
   Root cause: struct b1nix_tls_info used `unsigned long` (4 bytes on i686)
   while the kernel SYS_GET_TLS_INFO writes u64 fields. The 24-byte copyout
   overflowed the 12-byte userspace buffer and mis-paired the fields: filesz
   read the high half of the kernel's memsz (= 0), so build_thread_tls skipped
   the init-image copy. Fixed by making the struct fixed-width 64-bit (matches
   the kernel; 8 bytes on both arches). x86_64 was unaffected (long == u64).

2) displayd (and other long-running programs) #GP-faulted (err 0) on a global
   read under load -> SIGSEGV, cascading into graphics "(connect)" failures.
   Root cause: arch_check_and_deliver_signals re-enables interrupts, so the
   i686 return-to-user sequences ran with IF=1. A timer IRQ landing between the
   user-segment reload and the iret returns to kernel CS, so its own `testl $3`
   skips the user-segment restore and leaves %ds/%es = kernel 0x10; the iret to
   ring 3 then runs userspace with a DPL-0 data segment, and the next global
   access #GPs. Fixed with `cli` before the segment reload in syscall_entry.S,
   isr_common and irq_common, making segment-reload..iret atomic (iret restores
   the user IF from the saved eflags). long mode ignores segment limits/DPL, so
   x86_64 was immune.

Also: the exception dump now prints the faulting user ds/es/fs/gs (the CPU does
not push them); isr_common stashes them before reloading kernel selectors. This
is what pinpointed the corrupt %ds and aids any future segment bug.

Verified: x86 smoke 796/0/0; displayd survives host load ~14 (previously
crashed at ~4.5); full M29 pthread block green on i686.

+++ userspace/include/syscall.h
@@ -194,10 +194,16 @@ enum {
 /* PT_TLS template for the running image, returned by SYS_GET_TLS_INFO so the
  * libc can build a per-thread ELF TLS block in pthread_create (the kernel sets
  * up only the main thread's TLS at exec). memsz==0 => the binary has no TLS. */
+/* Fixed-width 64-bit fields to match the kernel's `u64 memsz, filesz, align`
+ * layout exactly. `unsigned long` is 4 bytes on i686, which mis-paired the
+ * fields against the kernel's 8-byte fields (userspace filesz read the high
+ * half of the kernel's memsz = 0) and overflowed this buffer by 12 bytes —
+ * breaking per-thread TLS init on the 32-bit port (thread-locals read 0
+ * instead of their .tdata image). 8 bytes on both i686 and x86_64. */
 struct b1nix_tls_info {
-  unsigned long memsz;  /* total TLS size (.tdata + .tbss) */
-  unsigned long filesz; /* initialised-image size (.tdata) */
-  unsigned long align;  /* TLS segment alignment (>=1) */
+  unsigned long long memsz;  /* total TLS size (.tdata + .tbss) */
+  unsigned long long filesz; /* initialised-image size (.tdata) */
+  unsigned long long align;  /* TLS segment alignment (>=1) */
 };
 
 /* Linux-compatible CLONE_* flag bits (subset honored by b1nix). */
